+++ docker/amster/git-sync.sh
@@ -8,9 +8,6 @@ GIT_ROOT=${GIT_ROOT:=/git/config}
 
 GIT_AUTOSAVE_BRANCH="${GIT_AUTOSAVE_BRANCH:-autosave-am}"
 
-export GIT_SSH_COMMAND="ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -i /etc/git-secret/ssh"
-
-
 cd "${GIT_ROOT}"
 
 # This configures git to ignore file mode changes.